+++ src/com/jec/juta/annotations/BeforeAll.ts
@@ -26,12 +26,12 @@ const DCM:DecoratorConnectorManager = DecoratorConnectorManager.getInstance();
 const CTXM:JcadContextManager = JcadContextManager.getInstance();
 
 /**
- * The <code>BeforeClass</code> decorator defines a method in the JUTA API which 
+ * The <code>BeforeAll</code> decorator defines a method in the JUTA API which 
  * is called one time before all tests run.
  *
  * ```javascript
-BeforeClass()
-public beforeClassMethod():void {
+BeforeAll()
+public beforeAllMethod():void {
   // Config here...
 }
 ```
@@ -39,7 +39,7 @@ public beforeClassMethod():void {
  * @param {AnnotatedMethodParams} params the parameters for the associated
  *                                       method.
  */
-export function BeforeClass(params?:AnnotatedMethodParams):Function {
+export function BeforeAll(params?:AnnotatedMethodParams):Function {
   
   return function(target:any, key:string,
                                        descriptor:PropertyDescriptor):Function {
@@ -49,8 +49,8 @@ export function BeforeClass(params?:AnnotatedMethodParams):Function {
     ////////////////////////////////////////////////////////////////////////////
 
     var ctx:JcadContext =
-                  CTXM.getContext(JutaConnectorRefs.BEFORE_CLASS_CONNECTOR_REF);
-    return DCM.getDecorator(JutaConnectorRefs.BEFORE_CLASS_CONNECTOR_REF, ctx)
+                    CTXM.getContext(JutaConnectorRefs.BEFORE_ALL_CONNECTOR_REF);
+    return DCM.getDecorator(JutaConnectorRefs.BEFORE_ALL_CONNECTOR_REF, ctx)
               .decorate(target, key, descriptor, params);
   }
 }