+++ src/com/jec/juta/index.ts
@@ -19,9 +19,10 @@ export {TestSuiteParams} from "./annotations/core/TestSuiteParams";
 //--> com/jec/juta/annotations
 export {After} from "./annotations/After";
 export {Async} from "./annotations/Async";
-export {AfterClass} from "./annotations/AfterClass";
+export {AfterAll} from "./annotations/AfterAll";
+//export {AfterClass} from "./annotations/AfterClass";
 export {Before} from "./annotations/Before";
-export {BeforeClass} from "./annotations/BeforeClass";
+export {BeforeAll} from "./annotations/BeforeAll";
 export {Test} from "./annotations/Test";
 export {TestSuite} from "./annotations/TestSuite";
 //--> com/jec/juta/exceptions