+++ src/com/jec/juta/utils/AnnotatedMethodType.ts
@@ -24,20 +24,30 @@ export enum AnnotatedMethodType {
   /**
    * Specifies the type for a <code>@BeforeClass</code> annotation.
    */
-  BEFORE_CLASS = 0,
+  //BEFORE_CLASS = 0,
 
   /**
    * Specifies the type for a <code>@AfterClass</code> annotation.
    */
-  AFTER_CLASS = 1,
+  //AFTER_CLASS = 1,
+
+  /**
+   * Specifies the type for a <code>@BeforeAll</code> annotation.
+   */
+  BEFORE_ALL = 2,
+
+  /**
+   * Specifies the type for a <code>@AfterAll</code> annotation.
+   */
+  AFTER_ALL = 3,
 
   /**
    * Specifies the type for a <code>@Before</code> annotation.
    */
-  BEFORE = 2,
+  BEFORE = 4,
 
   /**
    * Specifies the type for a <code>@After</code> annotation.
    */
-  AFTER = 3
+  AFTER = 5
 }