+++ utils/test-utils/classes/annotations/AfterClassTestClass.ts
@@ -1,8 +1,0 @@
-import {AfterClass} from "../../../../src/com/jec/juta/annotations/AfterClass";
-import * as params from "../../utilities/ParamUtils";
-
-export class AfterClassTestClass {
-
-  @AfterClass(params.ANNOTATED_METHOD_PARAMS)
-  public afterClassMethod():void { }
-}