+++ utils/test-utils/classes/annotations/BeforeAllTestClass.ts
@@ -1,0 +1,8 @@
+import {BeforeAll} from "../../../../src/com/jec/juta/annotations/BeforeAll";
+import * as params from "../../utilities/ParamUtils";
+
+export class BeforeAllTestClass {
+
+  @BeforeAll(params.ANNOTATED_METHOD_PARAMS)
+  public beforeAllMethod():void { }
+}