+++ utils/test-utils/classes/annotations/BeforeClassTestClass.ts
@@ -1,8 +1,0 @@
-import {BeforeClass} from "../../../../src/com/jec/juta/annotations/BeforeClass";
-import * as params from "../../utilities/ParamUtils";
-
-export class AfterClassTestClass {
-
-  @BeforeClass(params.ANNOTATED_METHOD_PARAMS)
-  public beforeClassMethod():void { }
-}