+++ utils/types/jec-juta.d.ts
@@ -8,5 +8,4 @@
 declare module "jec-juta" {
 
 /*PLACEHOLDER*/
-
 }